chore: Initial setup
Everything is all over the place right now, the stuff that's working is the dotfile management via PSDotFiles as well as a basic komorebi config. Bootstrapping is in a bit of chaos right now since installing applications/packages is not straight fowrard on Windows (unfortunately). Nix configuratio not even started.